minor fixes

+++ client/src/Components/FeedCardLeft2.jsx
@@ -1,35 +1,64 @@
 /* eslint-disable no-unused-vars */
 /* eslint-disable react/prop-types */
+import axios from "axios";
+import { useEffect, useState } from "react";
+import { useNavigate, useParams } from "react-router-dom";
 import { jwtDecode } from "jwt-decode";
-import BASE_API from '../api.js'
+import BASE_API from '../api.js';
+import TimeAgo from 'react-timeago';
 
-
-const FeedCard = ({ title, content, image, author, username, createdAt, time, onRemove }) => {
-
+const FeedCard = ({ title, content, image, author, username, createdAt, onRemove, myprofile }) => {
   const token = localStorage.getItem('token');
+  const {currAuthor} = useParams();
   const userData = token ? jwtDecode(token) : null;
+  const navigate = useNavigate();
+  const [ppUrl, setPpUrl] = useState('');
 
-
-  const timeAgo = (createdAt) => {
-    const now = new Date();
-    const createdAtDate = new Date(createdAt);
-    const diffTime = Math.abs(now - createdAtDate);
-    const diffDays = Math.floor(diffTime / (1000 * 60 * 60 * 24));
-
-    if (diffDays === 0) {
-      return "Today";
-    } else {
-      return `${diffDays} day${diffDays !== 1 ? 's' : ''} ago`;
-    }
-  };
+  useEffect(() => {
+    const fetchUserDetails = async () => {
+      try {
+        const response = await axios.get(`${BASE_API}/auth/getUserDetails`, {
+          headers: {
+            'Authorization': `Bearer ${author}`
+          }
+        });
+        setPpUrl(response.data.profilePicUrl);
+      } catch (error) {
+        console.error('Error fetching user details:', error);
+      }
+    };
+    fetchUserDetails();
+  }, [author]);
 
   const handleRemove = () => {
-    onRemove();
+    if (onRemove) {
+      onRemove();
+    }
   };
 
   return (
     <div className="w-full p-2 relative">
       <div className="border rounded-lg shadow bg-white dark:border-gray-700 h-auto flex flex-col overflow-y-auto myposts" style={{ maxHeight: '560px' }}>
+        {(myprofile === 0) && (
+          <div className="flex justify-between items-center w-80 my-2">
+            <div className="flex items-center cursor-pointer" onClick={() => { navigate(`/account/${author}`, { state: { userEmail: author } }) }}>
+
+              <img
+                className="w-8 h-8 rounded-full mx-2"
+                src={`${BASE_API}/profilePicLocation/${ppUrl}`}
+                alt={`${username}'s Profile`}
+              />
+
+              <p className="text-xs font-medium text-gray-700">
+                {username}
+              </p>
+            </div>
+            <div>
+              <p className="text-xs text-gray-400 right-2">
+                <TimeAgo date={createdAt} />
+              </p>
+            </div>
+          </div>)}
         {image && (
           <div className="relative" style={{ paddingBottom: '56.25%' }}>
             <div className="absolute inset-0 flex items-center justify-center">
